fixes https://github.com/winjs/winjs/issues/1300
Adds a new member function onto _CommandingSurface "takeFocus()" which will focus the first focusable element in the CommandingSurface's content element when called.

Adds a custom implementation of onTakeFocus() to the AppBar and ToolBar lightDismissisableElement which calls _CommandingSurface.takeFocus()

Consolidates the tets for AppBar/ToolBar/CommandingSurface into on folder, so it is easier to run tests for all 3 components at the same time.